Building workspaces api, small changes in workspaces sql schema
