+++ package.json
@@ -42,6 +42,7 @@
     "react-hook-form": "^7.48.2",
     "tailwind-merge": "^2.0.0",
     "tailwindcss-animate": "^1.0.7",
+    "winston": "^3.12.0",
     "zod": "^3.22.4"
   },
   "devDependencies": {
