+++ src/lib/mysql/schema.ts
@@ -42,7 +42,7 @@ export const workspaces = mysqlTable("workspaces", {
   workspaceOwner: int("workspace_owner")
     .references(() => users.id, { onDelete: "cascade" })
     .notNull(),
-  title: text("title"),
+  title: text("title").notNull(),
   icon: text("icon"),
   inTrash: boolean("in_trash"),
   logo: text("logo"),
